+++ CodiceFiscale.Tests/CodiceFiscaleTokenizerTests.cs
@@ -1,6 +1,0 @@
-namespace CodiceFiscale.Tests;
-
-public class CodiceFiscaleTokenizerTests
-{
-    
-}